Update favicon and refactor global styles; replace DM Sans with Satoshi font, adjust typography scale, and enhance layout consistency across components.
