+++ app/about/styles.css
@@ -1,5 +1,3 @@
-
-
 .about img {
   width: 100%;
   height: 100%;
@@ -30,12 +28,12 @@
   overflow: hidden;
 }
 
-.about .hero {
+.about .about-me {
   flex-direction: column;
   transition: background-color 0.3s ease;
 }
 
-.about .hero-header {
+.about .about-me-header {
   position: absolute;
   top: 35%;
   left: 50%;
@@ -57,13 +55,14 @@
   align-items: center;
   gap: 1rem;
   will-change: transform;
-  z-index: 2;
+  z-index: 10;
 }
 
 .animated-icon {
   flex: 1;
   aspect-ratio: 1;
   will-change: transform;
+  z-index: 10;
 }
 
 .animated-text {
@@ -99,7 +98,7 @@
     font-size: 1.1rem;
   }
 
-  .about .hero-header {
+  .about .about-me-header {
     top: 45%;
     width: 100%;
   }
@@ -112,6 +111,6 @@
 
 .duplicate-icon {
   position: absolute !important;
-  z-index: 1000;
+  z-index: 999;
   pointer-events: none;
 }
