+++ app/globals.css
@@ -1,11 +1,10 @@
 @import "tailwindcss";
 @import "tw-animate-css";
 
-
 :root {
   --dark: #1e1e1e;
   --light: #fefff8;
-  --accent: #ff3500;
+  --orange: #ff3500;
   --radius: 0.625rem;
   --background: oklch(1 0 0);
   --foreground: oklch(0.145 0 0);
@@ -86,8 +85,8 @@
 }
 
 body {
-  font-family: var(--font-dm-sans), sans-serif;
-  background-color: #FAFAFA;
+  font-family: var(--font-satoshi), sans-serif;
+  background-color: #fafafa;
   overflow-x: hidden;
   overflow-y: scroll; /* Force scrollbar to always be present */
 }
@@ -98,18 +97,171 @@ nav img {
   object-fit: cover;
 }
 
+/* ===============================================
+   TYPOGRAPHY SCALE SYSTEM
+   =============================================== */
+
+/* Headings */
 h1 {
-  text-transform: uppercase;
-  font-family: var(--font-anton), sans-serif;
-  font-size: 10rem;
-  font-weight: 400;
-  letter-spacing: -0.1rem;
+  font-size: clamp(3rem, 8vw, 8rem);
   line-height: 0.9;
+  letter-spacing: -0.04em;
+  margin-bottom: 2rem;
 }
 
-.anton-font {
-  font-family: var(--font-anton), sans-serif;
-  font-weight: 400;
+h2 {
+  font-size: clamp(2rem, 5vw, 4rem);
+  line-height: 1;
+  letter-spacing: -0.02em;
+  margin-bottom: 1.5rem;
+}
+
+h3 {
+  font-size: clamp(1.5rem, 3vw, 2.5rem);
+  line-height: 1.1;
+  letter-spacing: -0.015em;
+  margin-bottom: 1.25rem;
+}
+
+h4 {
+  font-size: clamp(1.25rem, 2vw, 1.75rem);
+  line-height: 1.2;
+  letter-spacing: -0.01em;
+  margin-bottom: 1rem;
+}
+
+h5 {
+  font-size: clamp(1.125rem, 1.5vw, 1.375rem);
+  line-height: 1.3;
+  letter-spacing: 0;
+  margin-bottom: 0.875rem;
+}
+
+h6 {
+  font-size: clamp(1rem, 1.25vw, 1.125rem);
+  line-height: 1.35;
+  letter-spacing: 0.01em;
+  margin-bottom: 0.75rem;
+}
+
+/* Body Text */
+p {
+  font-size: clamp(0.9rem, 1.1vw, 1.125rem);
+  line-height: 1.65;
+  letter-spacing: 0.005em;
+  margin-bottom: 1.5rem;
+}
+
+/* Lists */
+ul,
+ol {
+  font-size: clamp(0.9rem, 1.1vw, 1.125rem);
+  line-height: 1.65;
+  letter-spacing: 0.005em;
+  margin-bottom: 1.5rem;
+  padding-left: 1.5rem;
+}
+
+li {
+  margin-bottom: 0.5rem;
+}
+
+/* Links */
+a {
+  font-size: inherit;
+  line-height: inherit;
+  letter-spacing: 0.01em;
+}
+
+/* Small Text */
+small {
+  font-size: clamp(0.75rem, 0.9vw, 0.875rem);
+  line-height: 1.5;
+  letter-spacing: 0.02em;
+}
+
+/* Code and Preformatted */
+code {
+  font-size: clamp(0.8rem, 0.9vw, 0.9rem);
+  line-height: 1.6;
+  letter-spacing: 0.02em;
+}
+
+pre {
+  font-size: clamp(0.8rem, 0.9vw, 0.9rem);
+  line-height: 1.6;
+  letter-spacing: 0.02em;
+  margin-bottom: 1.5rem;
+  padding: 1rem;
+}
+
+/* Blockquote */
+blockquote {
+  font-size: clamp(1rem, 1.25vw, 1.25rem);
+  line-height: 1.6;
+  letter-spacing: 0.01em;
+  margin-bottom: 2rem;
+  padding-left: 2rem;
+}
+
+/* Strong and Emphasis */
+strong,
+b {
+  font-size: inherit;
+  line-height: inherit;
+  letter-spacing: inherit;
+}
+
+em,
+i {
+  font-size: inherit;
+  line-height: inherit;
+  letter-spacing: 0.005em;
+}
+
+/* Captions and Labels */
+figcaption {
+  font-size: clamp(0.75rem, 0.85vw, 0.8rem);
+  line-height: 1.5;
+  letter-spacing: 0.02em;
+  margin-top: 0.5rem;
+}
+
+label {
+  font-size: clamp(0.8rem, 0.9vw, 0.9rem);
+  line-height: 1.4;
+  letter-spacing: 0.02em;
+  margin-bottom: 0.5rem;
+}
+
+/* Form Elements */
+input,
+textarea,
+select {
+  font-size: clamp(0.9rem, 1vw, 1rem);
+  line-height: 1.5;
+  letter-spacing: 0.01em;
+}
+
+button {
+  font-size: clamp(0.8rem, 0.9vw, 0.9rem);
+  line-height: 1.4;
+  letter-spacing: 0.05em;
+}
+
+/* Table Elements */
+th {
+  font-size: clamp(0.8rem, 0.9vw, 0.9rem);
+  line-height: 1.4;
+  letter-spacing: 0.02em;
+  padding: 0.75rem;
+}
+
+td {
+  font-size: clamp(0.85rem, 0.95vw, 0.95rem);
+  line-height: 1.5;
+  letter-spacing: 0.01em;
+  padding: 0.75rem;
 }
 section {
   position: relative;
@@ -117,7 +269,6 @@ section {
   height: 100vh;
   overflow: hidden;
 }
-
 
 .menu p,
 .menu a {
@@ -129,51 +280,15 @@ section {
   user-select: none;
 }
 
-
-
-.hero h1 {
+.about-me h1 {
   width: 70%;
 }
 
 @media (max-width: 1000px) {
-  .hero h1 {
+  .about-me h1 {
     width: 100%;
     font-size: 4rem;
   }
-}
-
-.dark {
-  --background: oklch(0.145 0 0);
-  --foreground: oklch(0.985 0 0);
-  --card: oklch(0.205 0 0);
-  --card-foreground: oklch(0.985 0 0);
-  --popover: oklch(0.205 0 0);
-  --popover-foreground: oklch(0.985 0 0);
-  --primary: oklch(0.922 0 0);
-  --primary-foreground: oklch(0.205 0 0);
-  --secondary: oklch(0.269 0 0);
-  --secondary-foreground: oklch(0.985 0 0);
-  --muted: oklch(0.269 0 0);
-  --muted-foreground: oklch(0.708 0 0);
-  --accent: oklch(0.269 0 0);
-  --accent-foreground: oklch(0.985 0 0);
-  --destructive: oklch(0.704 0.191 22.216);
-  --border: oklch(1 0 0 / 10%);
-  --input: oklch(1 0 0 / 15%);
-  --ring: oklch(0.556 0 0);
-  --chart-1: oklch(0.488 0.243 264.376);
-  --chart-2: oklch(0.696 0.17 162.48);
-  --chart-3: oklch(0.769 0.188 70.08);
-  --chart-4: oklch(0.627 0.265 303.9);
-  --chart-5: oklch(0.645 0.246 16.439);
-  --sidebar: oklch(0.205 0 0);
-  --sidebar-foreground: oklch(0.985 0 0);
-  --sidebar-primary: oklch(0.488 0.243 264.376);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.269 0 0);
-  --sidebar-accent-foreground: oklch(0.985 0 0);
-  --sidebar-border: oklch(1 0 0 / 10%);
-  --sidebar-ring: oklch(0.556 0 0);
 }
 
 @layer base {
