+++ components/Footer/Footer.css
@@ -6,7 +6,7 @@
   border-radius: 20px;
   position: relative;
   overflow: hidden;
-  z-index: 1
+  z-index: 1;
 }
 
 .footer-content {
